add new Component PlacesCard for main & offer pages
