+++ src/types/image.ts
@@ -1,0 +1,1 @@
+export type CardsImageSize = 'small' | 'large';
